+++ components/tab-components/index.ts
@@ -1,2 +1,2 @@
-export { CategoryUl } from "./CategoryUl";
+export { CategoryUl } from "./TabUl";
 export { CategoryMenuItems } from "./TabItems";
